FIX: TransactionInputs are now sorted by Id during inflation.

+++ src/main/java/com/softwareverde/bitcoin/server/Configuration.java
@@ -36,7 +36,7 @@ public class Configuration {
         private Integer _maxPeerCount;
         private Integer _maxBlockQueueSize;
         private Integer _maxThreadCount;
-        private Integer _trustedBlockHeight;
+        private Long _trustedBlockHeight;
         private Long _maxMemoryByteCount;
         private Boolean _shouldSkipNetworking;
 
@@ -47,7 +47,7 @@ public class Configuration {
         public Integer getMaxPeerCount() { return _maxPeerCount; }
         public Integer getMaxBlockQueueSize() { return _maxBlockQueueSize; }
         public Integer getMaxThreadCount() { return _maxThreadCount; }
-        public Integer getTrustedBlockHeight() { return _trustedBlockHeight; }
+        public Long getTrustedBlockHeight() { return _trustedBlockHeight; }
         public Long getMaxMemoryByteCount() { return _maxMemoryByteCount; }
         public Boolean skipNetworking() { return _shouldSkipNetworking; }
     }
@@ -126,7 +126,7 @@ public class Configuration {
         _serverProperties._maxPeerCount = Util.parseInt(_properties.getProperty("bitcoin.maxPeerCount", "24"));
         _serverProperties._maxBlockQueueSize = Util.parseInt(_properties.getProperty("bitcoin.maxBlockQueueSize", "56"));
         _serverProperties._maxThreadCount = Util.parseInt(_properties.getProperty("bitcoin.maxThreadCount", "4"));
-        _serverProperties._trustedBlockHeight = Util.parseInt(_properties.getProperty("bitcoin.trustedBlockHeight", "0"));
+        _serverProperties._trustedBlockHeight = Util.parseLong(_properties.getProperty("bitcoin.trustedBlockHeight", "0"));
         _serverProperties._maxMemoryByteCount = Util.parseLong(_properties.getProperty("bitcoin.maxMemoryByteCount", String.valueOf(2L * ByteUtil.Unit.GIGABYTES)));
         _serverProperties._shouldSkipNetworking = Util.parseBool(_properties.getProperty("bitcoin.skipNetworking", "0"));
     }
